+++ lib/src/client/systems/events.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
-import 'dart:html';
 
 import 'package:damacreat/damacreat.dart';
 import 'package:damacreat_io/shared.dart';
@@ -9,105 +8,6 @@ import 'package:gamedev_helpers/gamedev_helpers.dart' hide Velocity;
 import 'package:damacreat_io/src/shared/components.dart';
 
 part 'events.g.dart';
-
-@Generate(
-  EntityProcessingSystem,
-  allOf: [
-    Controller,
-    Booster,
-  ],
-  manager: [
-    GameStateManager,
-    CameraManager,
-  ],
-)
-class ControllerSystem extends _$ControllerSystem {
-  WebSocketHandler _webSocketHandler;
-  CanvasElement canvas;
-  Point<num> offset;
-  bool useBooster = false;
-  int boosterFinger;
-
-  ControllerSystem(this.canvas, this._webSocketHandler);
-
-  @override
-  void initialize() {
-    super.initialize();
-    canvas.onMouseMove.listen((event) {
-      offset = event.offset;
-    });
-    canvas.onTouchMove.listen(_handleTouchEvent);
-    canvas.onTouchStart.listen((event) {
-      _handleTouchEvent(event);
-      event.preventDefault();
-    });
-    canvas.onTouchEnd.listen((event) {
-      for (final touch in event.changedTouches) {
-        if (touch.identifier == boosterFinger) {
-          useBooster = false;
-          boosterFinger = null;
-        }
-      }
-      event.preventDefault();
-    });
-    canvas.onMouseDown.listen((event) {
-      if (event.buttons & 1 == 1) {
-        useBooster = true;
-      }
-    });
-    canvas.onMouseUp.listen((event) {
-      if (event.buttons & 1 == 0) {
-        useBooster = false;
-      }
-    });
-  }
-
-  void _handleTouchEvent(TouchEvent event) {
-    final boosterOffset = Point<double>(boosterButtonCenterX.toDouble(),
-        cameraManager.clientHeight - boosterButtonCenterY.toDouble());
-    for (final touch in event.targetTouches) {
-      final touchOffset = touch.page;
-      if (boosterOffset.distanceTo(touchOffset) <= boosterButtonRadius) {
-        useBooster = true;
-        boosterFinger = touch.identifier;
-      } else {
-        offset = touchOffset;
-        if (boosterFinger == touch.identifier) {
-          useBooster = false;
-          boosterFinger = null;
-        }
-      }
-    }
-  }
-
-  @override
-  void processEntity(Entity entity) {
-    boosterMapper[entity].inUse = useBooster;
-    if (offset != null) {
-      final center = Point<num>(canvas.width / 2, canvas.height / 2);
-      final maxDistance = min(canvas.width / 3, canvas.height / 3);
-      final distance = center.distanceTo(offset);
-      final velocity =
-          ByteUtils.speedToByte(min(maxDistance, distance) / maxDistance);
-      final angle = ByteUtils.angleToByte(
-          tau + atan2(center.y - offset.y, offset.x - center.x));
-
-      final type = useBooster
-          ? MessageToServer.updateVelocityAndUseBooster
-          : MessageToServer.updateVelocity;
-      _webSocketHandler.sendData(Uint8ListWriter.clientToServer(type)
-        ..writeUint16(velocity)
-        ..writeUint16(angle));
-    } else if (useBooster) {
-      _webSocketHandler
-          .sendData(Uint8ListWriter.clientToServer(MessageToServer.useBooster));
-    }
-    offset = null;
-  }
-
-  @override
-  bool checkProcessing() => gameStateManager.state == GameState.playing;
-}
 
 @Generate(
   VoidEntitySystem,