+++ lib/src/ng/game_menu/game_menu.dart
@@ -2,6 +2,7 @@ import 'package:angular/angular.dart';
 import 'package:damacreat/damacreat.dart';
 import 'package:damacreat_io/src/ng/color_picker/color_picker.dart';
 import 'package:damacreat_io/src/ng/game_service.dart';
+import 'package:damacreat_io/src/shared/managers/controller_manager.dart';
 import 'package:damacreat_io/src/shared/managers/settings_manager.dart';
 import 'package:damacreat_io/src/version.dart';
 
@@ -18,12 +19,12 @@ class GameMenuComponent {
   GameService service;
   GameMenuComponent(this.service);
 
-  void joinGame(String nickname) {
-    service.joinGame(nickname);
+  void joinGame() {
+    service.joinGame();
   }
 
   void setTouchscreenDevice() {
-    settings.isTouchScreen = true;
+    controller.type = ControllerType.touch;
   }
 
   void toggleChangelog() {
@@ -32,6 +33,7 @@ class GameMenuComponent {
 
   int get maxLength => maxLengthNickname;
   SettingsManager get settings => service.settings;
+  ControllerManager get controller => service.controllerManager;
   bool get connected =>
       service.connectionState == ServerConnectionState.connected;
   bool get connecting =>
@@ -39,5 +41,6 @@ class GameMenuComponent {
   bool get connectionError =>
       service.connectionState == ServerConnectionState.error;
   String get version => packageVersion;
-  String get name => service.lastName;
+  String get nickname => service.nickname;
+  set nickname(String value) => service.nickname = value;
 }