+++ lib/src/ng/game_service.dart
@@ -5,6 +5,7 @@ import 'dart:math';
 import 'package:damacreat_io/app_component.dart';
 import 'package:damacreat_io/client.dart';
 import 'package:damacreat_io/src/client/web_socket_handler.dart';
+import 'package:damacreat_io/src/shared/managers/controller_manager.dart';
 import 'package:damacreat_io/src/shared/managers/game_state_manager.dart';
 import 'package:damacreat_io/src/shared/managers/settings_manager.dart';
 
@@ -16,13 +17,15 @@ class GameService {
   bool showChangelog = false;
   Object errorMessage;
   StackTrace stackTrace;
-  String lastName = '';
+  String nickname = '';
   int hue = 0;
+  final Set<int> gamepadIndices = Set<int>();
   final Random random = Random();
 
   final SettingsManager settings;
   final GameStateManager gameStateManager;
-  GameService(this.settings, this.gameStateManager);
+  final ControllerManager controllerManager;
+  GameService(this.settings, this.gameStateManager, this.controllerManager);
 
   Future<void> init() async {
     hue = random.nextInt(256);
@@ -37,7 +40,9 @@ class GameService {
       webSocket.onOpen.listen((openEvent) {
         connectionState = ServerConnectionState.connected;
         final webSocketHandler = WebSocketHandler(webSocket, debug: debug);
-        _game = Game(webSocketHandler, settings, gameStateManager)..start();
+        _game = Game(
+            webSocketHandler, settings, gameStateManager, controllerManager)
+          ..start();
         window.onBeforeUnload.listen((_) {
           webSocket.close();
         });
@@ -45,6 +50,22 @@ class GameService {
       webSocket.onError.listen((errorEvent) {
         connectionState = ServerConnectionState.error;
       });
+
+      if (window.navigator.getGamepads != null) {
+        final connectedGamepads = window.navigator.getGamepads();
+        for (var i = 0; i < connectedGamepads.length; i++) {
+          if (connectedGamepads[i] != null) {
+            gamepadIndices.add(i);
+          }
+        }
+      }
+      window.on['gamepadconnected'].cast<GamepadEvent>().listen((event) {
+        gamepadIndices.add(event.gamepad.index);
+      });
+      window.on['gamepaddisconnected'].cast<GamepadEvent>().listen((event) {
+        gamepadIndices.remove(event.gamepad.index);
+      });
+      _handleGamepads();
     }, onError: (errorMessage, stackTrace) {
       error = true;
       this.errorMessage = errorMessage;
@@ -52,9 +73,31 @@ class GameService {
     });
   }
 
-  void joinGame(String nickname) {
+  void _handleGamepads([_]) {
+    if (gameStateManager.state == GameState.menu && gamepadIndices.isNotEmpty) {
+      // window.navigator.getGamepads() needs to be used because otherwise
+      // pressed buttons will not be recogniced in chrome after a page refresh
+      final gamepads = window.navigator.getGamepads();
+      for (final index in gamepadIndices) {
+        final gamepad = gamepads[index];
+        if (gamepad.buttons[9].pressed) {
+          joinGame(gamepad: gamepad);
+          break;
+        }
+      }
+    }
+    window.requestAnimationFrame(_handleGamepads);
+  }
+
+  void joinGame({Gamepad gamepad}) {
     if (!error && menuVisible) {
-      lastName = nickname;
+      if (gamepad != null) {
+        controllerManager
+          ..type = ControllerType.gamepad
+          ..gamepadIndex = gamepad.index;
+      } else if (controllerManager.type == ControllerType.gamepad) {
+        controllerManager.type = ControllerType.mouse;
+      }
       _game.joinGame(hue ?? random.nextInt(256), nickname);
       gameStateManager.state = GameState.playing;
     }